Arreglada esfera xd falta que los ejes no se descoloreen

+++ Practica1-IG/ProyectosIG_x64_VS2019_FG/IG1App/Entity.h
@@ -192,7 +192,7 @@ public:
 
 class Esfera : public EntityWithIndexMesh {
 public:
-	explicit Esfera(GLdouble r, GLdouble p, GLdouble m, glm::fvec3 color);
+	explicit Esfera(GLdouble r, GLuint p, GLuint m, glm::fvec3 color);
 	~Esfera();
 	virtual void render(glm::dmat4 const& modelViewMat) const;
 	virtual void update() {};
